+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.0.2
 
 - Fix bug that would occur if Journal entries weren't in a folder of the same name as the adventure.
+- Delay calling `scenePackerReady` hook until after the canvas is ready.
 
 ## v1.0.1